tweaked how number of fetched days is counted
We look at the whole fetched dataset, before filtering out items past the scan window, to determine if we managed to fetch the whole window or not. This makes a different in situations where e.g. we want to fetch last 10 days, and someone only made one post 7 days ago; in that case, we fetch well beyond the 10-day window, but previously the returned dataset would only have that one post 7 days ago, so it would seem as if we only fetched 7 days back.